+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JZ4TTKETTTSHGJE2SH62VAQJ"
+        "01JZ5S0EFBQQ7HT4GDSJSPPA2G"
     )
     r"""The id of the resource"""