+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -59,5 +59,5 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-07-02T05:47:42.508Z")
+    created: Optional[datetime] = parse_datetime("2025-07-02T14:35:11.157Z")
     r"""The date and time the resource was created"""