+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -100,7 +100,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JZ4TTKEVMRZPN3C6NK3KZFG2"
+        "01JZ5S0EFCQ2GVZAZBS6W9CH5G"
     )
     r"""The id of the resource"""