+++ packages/orq-rc/RELEASES.md
@@ -1909,3 +1909,13 @@ Based on:
 - [python v3.9.0-rc.12] packages/orq-rc
 ### Releases
 - [PyPI v3.9.0-rc.12] https://pypi.org/project/orq-ai-sdk/3.9.0-rc.12 - packages/orq-rc
+
+## 2025-07-02 14:36:55
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.575.0 (2.647.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.9.0-rc.13] packages/orq-rc
+### Releases
+- [PyPI v3.9.0-rc.13] https://pypi.org/project/orq-ai-sdk/3.9.0-rc.13 - packages/orq-rc